fix bag and fix function settings

+++ textjokeditor.cpp
@@ -7,7 +7,7 @@
 TextJokEditor::TextJokEditor(QWidget *parent)
     : QMainWindow(parent)
     , ui(new Ui::TextJokEditor)
-     // добавление диалога в программу
+     // добавление диалога настроек в программу
     , m_SettingsEditor (new SettingsEditor(this))
 {
 
@@ -20,105 +20,157 @@ TextJokEditor::TextJokEditor(QWidget *parent)
 
     setMenuBar(m_menu);
     addToolBar(m_toolbar);
+    readSettings();
+    applySettings();
+
+
 
 
 
     //new
-//    connect(m_menu.getAction()[0], &QAction::triggered, this, &TextJokEditor::fileNew);
-//    connect(m_toolbar->getAction()[0], &QAction::triggered, this, &TextJokEditor::fileNew);
-    connect(m_toolbar->getAction()[0], &QAction::triggered, this, &TextJokEditor::fileOpen);
-    connect(m_menu->getAction()[0], &QAction::triggered, this, &TextJokEditor::fileOpen);
-//    //open
-//    connect(m_menu.getAction()[1], &QAction::triggered, this, &TextJokEditor::fileOpen);
+    connect(m_toolbar->getAction()[0], &QAction::triggered, this, &TextJokEditor::fileNew);
+    connect(m_menu->getAction()[0], &QAction::triggered, this, &TextJokEditor::fileNew);
+   //open
     connect(m_toolbar->getAction()[1], &QAction::triggered, this, &TextJokEditor::fileOpen);
     connect(m_menu->getAction()[1], &QAction::triggered, this, &TextJokEditor::fileOpen);
 
-//    //save
-//    connect(m_menu.getAction()[2], &QAction::triggered, this, &TextJokEditor::fileSave);
+    //save
     connect(m_toolbar->getAction()[2], &QAction::triggered, this, &TextJokEditor::fileSave);
     connect(m_menu->getAction()[2], &QAction::triggered, this, &TextJokEditor::fileSave);
 
-//    //quit
-    connect(m_menu->getAction()[3], &QAction::triggered, this, &QApplication::quit);
+    //saveAs
+    connect(m_menu->getAction()[3], &QAction::triggered, this, &TextJokEditor::fileSaveAs);
 
-//    //undo
-//    connect(m_menu.getAction()[4], &QAction::triggered, this, &TextJokEditor::textUndo);
-    connect(m_menu->getAction()[3], &QAction::triggered, this, &TextJokEditor::textUndo);
-    connect(m_menu->getAction()[4], &QAction::triggered, this, &TextJokEditor::textUndo);
+    //quit
+    connect(m_menu->getAction()[4], &QAction::triggered, this, &QApplication::quit);
 
-//    //redo
-//    connect(m_menu.getAction()[5], &QAction::triggered, this, &TextJokEditor::textRedo);
+    //undo
+    connect(m_toolbar->getAction()[3], &QAction::triggered, this, &TextJokEditor::textUndo);
+    connect(m_menu->getAction()[5], &QAction::triggered, this, &TextJokEditor::textUndo);
+
+    //redo
     connect(m_toolbar->getAction()[4], &QAction::triggered, this, &TextJokEditor::textRedo);
-    connect(m_menu->getAction()[5], &QAction::triggered, this, &TextJokEditor::textRedo);
+    connect(m_menu->getAction()[6], &QAction::triggered, this, &TextJokEditor::textRedo);
 
-//    //cut
-//    connect(m_menu.getAction()[6], &QAction::triggered, this, &TextJokEditor::textCut);
+    //cut
     connect(m_toolbar->getAction()[5], &QAction::triggered, this, &TextJokEditor::textCut);
-    connect(m_menu->getAction()[6], &QAction::triggered, this, &TextJokEditor::textCut);
+    connect(m_menu->getAction()[7], &QAction::triggered, this, &TextJokEditor::textCut);
 
-//    //copy
-//    connect(m_menu.getAction()[7], &QAction::triggered, this, &TextJokEditor::textCopy);
+    //copy
    connect(m_toolbar->getAction()[6], &QAction::triggered, this, &TextJokEditor::textCopy);
-   connect(m_menu->getAction()[7], &QAction::triggered, this, &TextJokEditor::textCopy);
+   connect(m_menu->getAction()[8], &QAction::triggered, this, &TextJokEditor::textCopy);
 
-//    //paste
-//    connect(m_menu.getAction()[8], &QAction::triggered, this, &TextJokEditor::textCopy);
+    //paste
     connect(m_toolbar->getAction()[7], &QAction::triggered, this, &TextJokEditor::textPaste);
-    connect(m_menu->getAction()[8], &QAction::triggered, this, &TextJokEditor::textCopy);
+    connect(m_menu->getAction()[9], &QAction::triggered, this, &TextJokEditor::textCopy);
 
 //    //подключение сигнала настроек к слоту отображение настроек
-//    connect(m_menu.getAction()[9], &QAction::triggered, this, &TextJokEditor::showPreferencesDialog);
+    connect(m_menu->getAction()[10], &QAction::triggered, this, &TextJokEditor::showPreferencesDialog);
 
-
-    //connect(m_SettingsEditor,SIGNAL(accepted()), this, SLOT(slotPreferencesAccepted()));
     connect(m_SettingsEditor, &QDialog::accepted, this, &TextJokEditor::slotPreferencesAccepted);
 
     //подключение сигнала настроек к слоту отображение настроек
-    connect(ui->textField, SIGNAL(textChanged()), this, SLOT(countWS()));
+    connect(ui->textField, &QPlainTextEdit::textChanged, this, &TextJokEditor::countWS);
+
+    connect(ui->textField, &QPlainTextEdit::textChanged, this, &TextJokEditor::saveDocument );
 
 }
+
+//создание пустого файла с проверкой на сохранение файла
+void TextJokEditor::fileNew()
+{
+    //проверка на изменения файла
+    if (ui->textField->toPlainText().isEmpty())
+    {
+        m_textData.clear();
+        ui->textField->setPlainText(m_textData);
+        setWindowTitle("TextJokEditor") ;
+        m_isNewFile = true;
+
+    }else
+    {
+        if(m_isSaveDocument == false)
+        {
+            fileSave();
+//            if(m_isSaveDocument == false)
+//            {
+//                return;
+//            }
+            m_textData.clear();
+            ui->textField->setPlainText(m_textData);
+            setWindowTitle("TextJokEditor") ;
+            m_isNewFile = true;
+        }
+    }
+}
+
 //слот открытия файла
 void TextJokEditor::fileOpen(){
 
-    m_fileNameOpen = QFileDialog::getOpenFileName(this, "Open file ...", QDir::homePath(), "*.txt;; All files (*.*)");
+    if (m_isNewFile == true)
+    {
+        m_fileNameOpen = QFileDialog::getOpenFileName(this, "Open file ...", QDir::homePath(), "*.txt;; All files (*.*)");
 
-    //в случае пересохранении текстового файла
-    m_fileNameSave = m_fileNameOpen;
-
-    if (m_fileNameOpen.isEmpty())
-        return;
-
-    QFile file(m_fileNameOpen);
-
-
-    updateTitle(file.fileName());
-
-    if(file.exists()){
-        file.open(QIODevice::ReadOnly);
-
-        QTextStream in(&file);
-
-        m_textData.clear();
-        while (!in.atEnd()) {
-
-            m_textData.push_back(in.read(1));
-
+        if (m_fileNameOpen.isEmpty())
+        {
+            m_isOpenDocument = false;
+            return;
         }
-        ui->textField->setPlainText(m_textData);
+        m_isOpenDocument = true;
+        //в случае пересохранении текстового файла
+        m_fileNameSave = m_fileNameOpen;
+
+        if (m_fileNameOpen.isEmpty())
+            return;
+
+        QFile file(m_fileNameOpen);
+        updateTitle(file.fileName());
+
+        if(file.exists())
+        {
+
+            file.open(QIODevice::ReadWrite);
+            QTextStream in(&file);
+            m_textData.clear();
+
+            while (!in.atEnd())
+            {
+
+                m_textData.push_back(in.read(1));
+
+            }
+
+            ui->textField->setPlainText(m_textData);
+
+            m_isNewFile = false;
+
+        }else
+        {
+            QMessageBox::warning(this, "File not found", "File not found");
+        }
+    }else
+
+    //if (m_isNewFile == false)
+    {
+
+        //проверка на сохраненный документ
+        if (m_isSaveDocument == false)
+        {
+            fileSave();
+        }
+        m_isNewFile = true;
+        fileOpen();
 
 
     }
-    else {
-        QMessageBox::warning(this, "File not found", "File not found");
-    }
 
-    //QMessageBox::information(this, "Text", "test");
+
+    //QMessageBox::warning(this, "Warning", "Something wrong....try again");
+
 
 }
 
-
 //запись в файл
-
 void openFileWrite(QFile &file, QString &textData, QWidget *wid){
 
     file.open(QIODevice::Text | QIODevice::WriteOnly);
@@ -140,35 +192,44 @@ void TextJokEditor::fileSave()
 
     m_textData = ui->textField->toPlainText();
 
-    if (m_fileNameSave.isEmpty()){
-        QString m_fileNameSave = QFileDialog::getSaveFileName(this, "Save file...", QDir::homePath(), tr (" *.txt;; All files (*.*)" )) ;
-
-        if(m_fileNameSave.isEmpty())
-            return;
-
-        QFile file(m_fileNameSave);
-
-        openFileWrite(file, m_textData, this);
+    if (m_fileNameSave.isEmpty())
+    {
+        fileSaveAs();
+    }else
+    {
+        auto answer = QMessageBox::question(this, "Save?", "Аre you sure you want to save the document?");
+        if (answer == QMessageBox::Yes)
+        {
+            QFile file(m_fileNameSave);
+            file.remove();
+            openFileWrite(file, m_textData, this);
+            m_isSaveDocument = true;
+        }
+        if (answer == QMessageBox::No)
+        {
+            m_isSaveDocument = false;
+        }
     }
-    else{
 
-        QFile file(m_fileNameSave);
-        file.remove();
-        openFileWrite(file,m_textData, this);
-    }
+
 }
 
-//создание пустого файла с проверкой на сохранение файла
-void TextJokEditor::fileNew()
+void TextJokEditor::fileSaveAs()
 {
-    //проверка на изменения файла
-    if (true){
-        m_textData.clear();
-        ui->textField->setPlainText(m_textData);
+    m_textData = ui->textField->toPlainText();
 
-    }else {
-        fileSave();
+    QString m_fileNameSave = QFileDialog::getSaveFileName(this, "Save file...", QDir::homePath(), tr (" *.txt;; All files (*.*)" )) ;
+
+    if(m_fileNameSave.isEmpty())
+    {
+        m_isSaveDocument = false;
+        return;
     }
+
+    QFile file(m_fileNameSave);
+    openFileWrite(file, m_textData, this);
+    m_isSaveDocument = true;
+
 }
 
 //стандртые слоты при работе с текстом
@@ -199,9 +260,7 @@ void TextJokEditor::textRedo()
 //вывод в statusbar
 void TextJokEditor::outPutStatusBar(int first, int second)
 {
-
-    //ui->statusbar->showMessage(out.arg(first).arg(second));
-    label->setText(out.arg(first).arg(second));
+    label->setText(m_out.arg(first).arg(second));
     ui->statusbar->addWidget(label);
 }
 
@@ -211,23 +270,28 @@ void TextJokEditor::countWS()
     int countWord;
     int countSymbol = 0;
 
-    //countWord = ui->textField->toPlainText().split((" "), Qt::SkipEmptyParts).count();
-    //countWord += ui->textField->toPlainText().split(("\n"), Qt::SkipEmptyParts).count() - 1;
+    countWord = ui->textField->toPlainText().split((" "), QString::SkipEmptyParts).count();
+    countWord += ui->textField->toPlainText().split(("\n"), QString::SkipEmptyParts).count() - 1;
     if (countWord < 0)
         countWord = 0;
     countSymbol = ui->textField->toPlainText().length();
-
     outPutStatusBar(countWord,countSymbol);
 
+}
+
+void TextJokEditor::saveDocument()
+{
+    m_isSaveDocument = false;
 }
 
 //изменения титульника файла
 void TextJokEditor::updateTitle(QString fileName)
 {
-    QString title = QString("TextJokEditor - %1").arg(fileName); //устанавливаем заголовок окна
+    QString title = QString("TextJokEditor - %1").arg(fileName);
+    //устанавливаем заголовок окна
     setWindowTitle( title ) ;
 }
-//https://qthelper.ru/%D1%81%D0%BE%D1%85%D1%80%D0%B0%D0%BD%D0%B5%D0%BD%D0%B8%D0%B5-%D0%BD%D0%B0%D1%81%D1%82%D1%80%D0%BE%D0%B5%D0%BA-qsettings/
+
 
 //чтение сохраненных настроек
 void TextJokEditor::readSettings()
@@ -235,7 +299,8 @@ void TextJokEditor::readSettings()
     //Указываем,где хранились настройки.
     //QSettings::NativeFormat — в формате определенном системой
     //QSettings::UserScope — настройки для каждого пользователя отдельно.
-    QSettings Settings(QSettings::NativeFormat, QSettings::UserScope,"",qApp->applicationName());
+    //QSettings Settings(QSettings::NativeFormat, QSettings::UserScope,"",qApp->applicationName());
+    QSettings Settings("settings.conf", QSettings::IniFormat);
     //Открываем группу настроек
     Settings.beginGroup("SETTINGS_GROUP_VIEW");
     //Читаемнастройки
@@ -248,7 +313,7 @@ void TextJokEditor::readSettings()
 //запись настроек
 void TextJokEditor::writeSettings()
 {
-    QSettings Settings(QSettings::NativeFormat,QSettings::UserScope,"",qApp->applicationName());
+    QSettings Settings("settings.conf", QSettings::IniFormat);
     //Открываем группу настроек
     Settings.beginGroup("SETTINGS_GROUP_VIEW");
     //Записываем настройки
@@ -262,7 +327,7 @@ void TextJokEditor::applySettings()
 {
     //Читаем настройки установленые в диалоге и применяем их
     ui->statusbar->setVisible(m_SettingsEditor->isShowStatusBar());
-    //m_toolbar->setVisible(m_SettingsEditor->isShowToolBar());
+    m_toolbar->setVisible(m_SettingsEditor->isShowToolBar());
 }
 
 TextJokEditor::~TextJokEditor()
